Displace get_reader_class and list_writers
