+++ asreview/io/__init__.py
@@ -21,3 +21,5 @@ from asreview.io.paper_record import PaperRecord
 from asreview.io.ris_reader import RISReader
 from asreview.io.ris_writer import RISWriter
 from asreview.io.tsv_writer import TSVWriter
+from asreview.io.utils import get_reader_class
+from asreview.io.utils import list_writers
